Fix error message & move common dependency location
