+++ src/mock/v2.0/intent-c.ts
@@ -5,7 +5,7 @@ import {
 import { DesktopAgent } from "fdc3_2_0/dist/api/DesktopAgent";
 import { sendContextToTests } from "../v2.0/mock-functions";
 import { ContextToSend } from "./general";
-import { MockAppContext } from "../../test/v2.0/advanced/open-support-2.0";
+import { MockAppContext } from "../../test/common/open-control";
 declare let fdc3: DesktopAgent;
 onFdc3Ready().then(async () => {
   await closeWindowOnCompletion();
